triangle class created, WIP
I first tried a triangle class with rounded corners, but realized that it just didn't look good because of how small I would need the triangle to be for the triangle toggle button.

I decided on a triangle with no rounded corners. Instead of a circle thumb like in normal toggle buttons, my thumb will be a smaller triangle that will fit just inside one corner of the larger triangle. This is all still a WIP and this commit contains many errors and such.