WIP - triangle button can be toggled
The theming is not yet set up to reflect the selected option.
For some reason, the triangle gets bigger when the top option is selected.
The thumb is too large.
The color of the thumb (at least for the pastel theme) does not have enough contrast to the triangle button.
Still need to create the header and place the theme triangle toggle in there.