WIP -- triangle size changing
The thumb of the button is now the correct size -- even bigger than before, unexpectedly. I thought I would need to make it smaller, but it turns out that the thumb just needed to sit deeper inside the larger triangle. After that, a larger thumb made more sense.
The triangle toggle button itself no longer changes size when the length of the label changes when label is TOP or BOTTOM.
Button is not currently visible when the label is to the left or right.