Fixed button disappearing when label horizontal
Before, the label being on the left or right would cause issues where the button would disappear. I changed around the css so that the label has a fixed container size when in that orientation and truncates when it is too long. This also fixed the button moving when the label is in this configurationa and changes in size.